Improved coordinate transformations
- Integration happens in inertial frame (i.e., barycentric/center of
mass is the reference point) but additional accelerations are computed
in heliocentric frame (i.e., the star is the reference point),
conversion should take place carefully.
- Additional acceleration due to general relativy as implemented in
Anderson1975 and Newhall1983 are computed in inertial frame, thus they
have to be transformed to heliocentric to match the rest of additional
forces.
- WHFAST converts from/to inertial to/from an alternative coordinate
system that can be Jacobi, WHDS or DemocraticHeliocentric
- Simplified WHFAST json, now there is no need to clone universe.