Necessary functions to calculate migration timescale were added
The migration timescale calculation is done in a public function in
tools.rs. It needs other functions which are also in tools.rs like
calculate_disk_surface_density.
In constants: added the Boltzmann constant and the mass of the
hydrogen atom in the unit of M_SUN, AU and DAY.
In universe: added a line to be able to use the function
calculate_migration_timescale