migration_timescale is an output and removed the 2 in migration force
In order to test there is no weird business with the migration timescale
(is it negative somehow sometimes?), migration_timescale was added as an
output.
Also, there was a 2 in the expression of the migration force of
Alibert+13, there is no real reason for that. It was removed which makes
the new expression consistent with Mandell+07, Fogg&Nelson07.